refacto(game): remove easel ref and move state from easel to the game
In order to follow the React way to go, I move the logic from the easel to the game.
"React is all about one-way data flow down the component hierarchy"

The state of the easel impact both, the board and the easel component.
So putting the logic in the game component is more correct.