refacto(game): move state from board to the game
In order to follow the React way to go, I move the logic from the board to the game.
"React is all about one-way data flow down the component hierarchy"

The state of the board impact both, the board and the easel component.
So putting the logic in the game component is more correct.