fix(score): never count one letter word
One letter word was counted on the column direction when only one letter was played
whereas it was possible to already count a word for several letter on the line.
It results to a bad round score as one letter word must not be counted.

A complexity was present to handle the case of one letter played on the first round.
As it should never occur, I decided to simplify the calculation returning 0 when
only one letter is played at the first round.