feat(player): display the round scores separately of the sum
Instead of sum each round score with the previous sum,
we display the score of each round and add a sum at the end.

It will fix the round score comparaison on duplicate mode in the isNotWinner() function,
as the total scores from the beginning were compared instead of the round scores.